\name{collapse}
\alias{collapse}
\title{A simplied version of paste.}
\usage{collapse(...)}
\description{A simplied version of paste.}
\details{\code{collapse} A simplied version of paste.}
\value{A string of the values combined.}
\author{Shane Conway \email{shane.conway@gmail.com}}
\arguments{\item{...}{Comma separated list of values.}}
\examples{collapse("a", "b", "c")}
