\name{esse}
\alias{esse}
\title{Checks whether something exists.}
\usage{esse(x)}
\description{Checks whether something exists.}
\details{\code{esse} Checks whether something exists.}
\value{A logical value.}
\author{Shane Conway \email{shane.conway@gmail.com}}
\arguments{\item{x}{The object to be checked.}}
