\name{field.exists}
\alias{field.exists}
\title{Checks whether a field exists in a dataset.}
\usage{field.exists(field, data)}
\description{Checks whether a field exists in a dataset.}
\details{\code{field.exists} Checks whether a field exists in a dataset.}
\value{A logical value.}
\author{Shane Conway \email{shane.conway@gmail.com}}
\arguments{\item{field}{The field name.}
\item{data}{The data to be checked.}}
