\name{getHead}
\alias{getHead}
\title{Return the head html.}
\usage{getHead(title="", protovis.path=PROTOVIS.PATH)}
\description{Return the head html.}
\details{\code{getHead} Return the head html.}
\value{HTML.}
\author{Shane Conway \email{shane.conway@gmail.com}}
\arguments{\item{title}{The title of the page.}
\item{protovis.path}{The path (url) for protovis.}}
