\name{is.webvis.flat}
\alias{is.webvis.flat}
\title{Checks whether something is a webvis...}
\usage{is.webvis.flat(x)}
\description{Checks whether something is a webvis.flat object.}
\details{\code{is.webvis.flat} Checks whether something is a webvis.flat object.}
\value{A logical value.}
\author{Shane Conway \email{shane.conway@gmail.com}}
\arguments{\item{x}{The object to be checked.}}
