\name{is.webvis.param}
\alias{is.webvis.param}
\title{Checks whether something is a webvis...}
\usage{is.webvis.param(x)}
\description{Checks whether something is a webvis.param object.}
\details{\code{is.webvis.param} Checks whether something is a webvis.param object.}
\value{A logical value.}
\author{Shane Conway \email{shane.conway@gmail.com}}
\arguments{\item{x}{The object to be checked.}}
