\name{is.webvis}
\alias{is.webvis}
\title{Checks whether something is a webvis object.}
\usage{is.webvis(x)}
\description{Checks whether something is a webvis object.}
\details{\code{is.webvis} Checks whether something is a webvis object.}
\value{A logical value.}
\author{Shane Conway \email{shane.conway@gmail.com}}
\arguments{\item{x}{The object to be checked.}}
