\name{plot.webvis}
\alias{plot.webvis}
\title{Simplified plot function for web vis plots.}
\usage{plot.webvis(data, type="bar", width=500, height=500, ...)}
\description{Simplified plot function for web vis plots.}
\details{\code{plot.webvis} Simplified plot function for web vis plots}
\value{Opens a plot.}
\keyword{graphics}
\author{Shane Conway \email{shane.conway@gmail.com}}
\references{\url{http://vis.stanford.edu/protovis/}}
\seealso{\code{\link{new.webvis}} that creates the webvis object.}
\arguments{\item{data}{The webvis object containing the visualization.}
\item{type}{The type of plot.  Can be "bar", "line", "area", "pie", "dot", or "..."}
\item{width}{The width of the panel in pixels.}
\item{height}{The width of the panel in pixels.}}
