\name{pv.area}
\alias{pv.area}
\title{pv.area}
\usage{pv.area(wv, data, y.name="y", x.name="x", bottom=0, height, left, right, bar.width, line.width, stroke.style, segmented=(!missing(line.width) ||
    field.exists(field = "width", data = data)), interpolate, fill.style, x.padding=(wv$width)/50, y.padding=(wv$height)/50, xmin, xmax, ymin, ymax,
    scale="linear", anchor)}
