\name{pv.bar}
\alias{pv.bar}
\title{pv.bar}
\usage{pv.bar(wv, data, y.name="y", x.name="x", bottom=0, height, left, right,
    bar.width, line.width, stroke.style,
    segmented=(!missing(line.width) || field.exists(field = "width",
    data = data)), fill.style, x.padding=(wv$width)/50,
    y.padding=(wv$height)/50, xmin, xmax, ymin=y.padding, ymax,
    scale="linear", anchor)}
