\name{pv.data}
\alias{pv.data}
\title{Add a panel to the visualization.}
\usage{pv.data(data, quote=FALSE)}
\description{Add a panel to the visualization.}
\details{\code{protovis.data} Adds a panel to the visualization}
\value{A wv object.}
\keyword{graphics}
\author{Shane Conway \email{shane.conway@gmail.com}}
\references{\url{http://vis.stanford.edu/protovis/}}
\seealso{\code{\link{new.webvis}} that creates the webvis object.}
\arguments{\item{data}{The webvis object containing the visualization.}}
