\name{pv.dot}
\alias{pv.dot}
\title{wv <- new...}
\usage{pv.dot(left.name="x", bottom.name="y", bottom.scale=paste("linear",
    bottom.name, "y", sep = "."), left.scale=paste("linear", left.name,
    "x", sep = "."), ...)}
\description{wv <- new.webvis(width=150, height=150)
render.webvis(wv + pv.dot(wv=wv, data=c(1, 1.2, 1.7, 1.5, .7, .5, .2), ymax=140, xmax=140, xmin=10, ymin=10, scale.min=0))}
