\name{pv.panel}
\alias{pv.panel}
\title{A protovis panal.}
\usage{pv.panel(wv, data, width=300, height=200, left, right, bottom, top)}
\description{A protovis panal.}
\details{\code{pv.panel} A protovis panal.}
\value{The HTML output}
\keyword{graphics}
\author{Shane Conway \email{shane.conway@gmail.com}}
\references{\url{http://vis.stanford.edu/protovis/}}
\arguments{\item{param}{A webvis param object from pv.param().}
\item{wv}{A webvis object.}
\item{data}{A dataset.}}
