\name{render.webvis}
\alias{render.webvis}
\title{Create the final visualization from the webvis object.}
\usage{render.webvis(wv, vis.name, file.name, title="", protovis.path)}
\description{Create the final visualization from the webvis object.}
\details{\code{render.webvis} Renders the visualization from the webvis object.}
\value{The path to the output visualization.}
\keyword{hplot}
\author{Shane Conway \email{shane.conway@gmail.com}}
\references{\url{http://vis.stanford.edu/protovis/}}
\seealso{\code{\link{new.webvis}} that creates the webvis object.}
\arguments{\item{wv}{The webvis object containing the visualization.}
\item{vis.name}{The file name of the output HTML.}
\item{path}{The file path to the HTML file.}
\item{file.name}{The file path to the HTML file.}
\item{title}{The file path to the HTML file.}
\item{protovis.path}{The file path to the HTML file.}}
