\name{Rwebvis-package}
\title{Pulls data from NY Times API.}
\description{Pull data from the NY Times API}
\details{\tabular{ll}{
Package: \tab nytR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-12-23\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab no\cr
}

Pulls congressional data from NY Times API.  Currently only exposes the Congress API.  Requires an API key from \url{http://developer.nytimes.com/}.}
\alias{Rwebvis}
\docType{package}
\author{Shane Conway \email{shane.conway@gmail.com}}
\references{\url{http://developer.nytimes.com/}}
\keyword{package}
