\name{unfold.webvis}
\alias{unfold.webvis}
\title{Create the final visualization from the webvis object.}
\usage{unfold.webvis(wv, name="vis", parent)}
\description{Create the final visualization from the webvis object.}
\details{\code{unfold.webvis} Renders the visualization from the webvis object.}
\value{A wv object.}
\keyword{graphics}
\author{Shane Conway \email{shane.conway@gmail.com}}
\references{\url{http://vis.stanford.edu/protovis/}}
\seealso{\code{\link{new.webvis}} that creates the webvis object.}
\arguments{\item{wv}{The webvis object containing the visualization.}
\item{vis.name}{The file name of the output HTML.}
\item{path}{The file path to the HTML file.}
\item{file.name}{The file path to the HTML file.}
\item{title}{The file path to the HTML file.}
\item{protovis.path}{The file path to the HTML file.}}
