\name{webvis-package}
\title{Web graphics for R.}
\description{Provides web graphics for R by wrapping visualization API's.  Currently supports Protovis.}
\details{\tabular{ll}{
Package: \tab webvis\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2010-03-22\cr
License: \tab BSD (>= 2)\cr
LazyLoad: \tab no\cr
}

Uses Protovis to provide web graphics for R.  Package is still under active development and shouldn't be considered
stable until version 0.1.  Currently uses a web browser to process JavaScript, although future version will process 
JavaScript directly and return the SVG output.}
\alias{webvis}
\docType{package}
\author{Shane Conway \email{shane.conway@gmail.com}}
\references{\url{http://vis.stanford.edu/protovis}}
\keyword{package}
