\name{"+.webvis"}
\alias{+.webvis}
\title{Add a new layers to a visualization.}
\usage{"+.webvis"(parent, child)}
\description{Add a new layers to a visualization.}
\details{\code{+.webvis} Add a new layers to a visualization.}
\value{A webvis object.}
\keyword{hplot}
\author{Shane Conway \email{shane.conway@gmail.com}}
\references{\url{http://vis.stanford.edu/protovis/}}
\arguments{\item{parent}{The root node.}
\item{child}{The leaf node.}}
