\name{webvis.theme}
\alias{webvis.theme}
\title{Creates a theme to be used by a webvis chart object.}
\usage{webvis.theme(background, font, colors, grid)}
\description{Creates a theme to be used by a webvis chart object.}
\details{\code{webvis.theme} Creates a theme to be used by a webvis chart object.  Not currently used anywhere.}
\value{A webvis.theme object.}
\author{Shane Conway \email{shane.conway@gmail.com}}
\references{\url{http://vis.stanford.edu/protovis/}}
\arguments{\item{background}{The background color.}
\item{font}{The found used for any text.}
\item{colors}{The color spectrum.}
\item{grid}{The grid colors.}}
