\name{pv.data}
\alias{pv.data}
\title{Converts R data into Protovis data.}
\usage{pv.data(data, quote=FALSE)}
\description{Converts R data into Protovis data.}
\details{\code{pv.data} Converts R data into Protovis data.}
\value{A protovis data object.}
\keyword{hplot}
\author{Shane Conway \email{shane.conway@gmail.com}}
\references{\url{http://vis.stanford.edu/protovis/}}
\arguments{\item{data}{An R data object.}
\item{quote}{Whether characters should be quoted.}}
